# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Cetacean API
  version: "1.0"
  description: |
    Observability and management API for Docker Swarm Mode clusters.

    Uses content negotiation — same URLs serve JSON, HTML (SPA), or SSE based on
    the Accept header. Append `.json` or `.html` to any path to force a content type.

    All collection responses include JSON-LD `@context` and `@type` fields.
    Detail responses include `@context`, `@id`, and `@type`.
    Error responses follow RFC 9457 (Problem Details).

    Every response (except `/-/` meta endpoints) includes:
    - `Link: </api>; rel="service-desc"` — pointer to this spec
    - `Link: </api/context.jsonld>; rel="describedby"` — JSON-LD context
    - `Vary: Accept` — content negotiation indicator
servers:
  - url: /
security:
  - cookieAuth: []
  - bearerAuth: []
  - {} # Allow unauthenticated (none mode)
...
paths:
...
  /configs:
    get:
      operationId: listConfigs
      tags: [Configs]
      summary: List configs
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created, updated]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated config list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createConfig
      tags: [Configs]
      summary: Create a config
      description: Creates a new config in the swarm. Requires operations level 2.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, data]
              properties:
                name:
                  type: string
                  description: Config name
                data:
                  type: string
                  format: byte
                  description: Config data (base64-encoded)
      responses:
        "201":
          description: Config created
          headers:
            Location:
              schema:
                type: string
              description: Path to the created config
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      config:
                        type: object
                        description: Docker swarm.Config object (data is base64-encoded)
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Operations level too low
        "409":
          description: Config name conflict
  /configs/{id}:
    get:
      operationId: getConfig
      tags: [Configs]
      summary: Config detail
      description: |
        Returns the config with its data (base64-encoded) and
        a list of services using this config.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Config detail with cross-referenced services
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      config:
                        type: object
                        description: Docker swarm.Config object (data is base64-encoded)
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeConfig
      tags: [Configs]
      summary: Remove a config
      description: Removes a config from the swarm. Fails if the config is in use by a service. Requires operations level 3.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "204":
          description: Config removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /configs/{id}/labels:
    get:
      operationId: getConfigLabels
      tags: [Configs]
      summary: Get config labels
      description: Returns the current labels of the config as a key/value map.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      responses:
        "200":
          description: Config labels
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchConfigLabels
      tags: [Configs]
      summary: Patch config labels
      description: |
        Applies a JSON Patch (RFC 6902) to the config labels.
        Requires `Content-Type: application/json-patch+json`.
      parameters:
        - $ref: "#/components/parameters/resourceId"
      requestBody:
        required: true
        content:
          application/json-patch+json:
            schema:
              $ref: "#/components/schemas/JSONPatch"
      responses:
        "200":
          description: Updated config labels
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"

  # ── Secrets ──────────────────────────────────────────────────────────
...
  /volumes:
    get:
      operationId: listVolumes
      tags: [Volumes]
      summary: List volumes
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/filter"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, driver, scope]
        - $ref: "#/components/parameters/dir"
      responses:
        "200":
          description: Paginated volume list
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Link:
              $ref: "#/components/headers/Link"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolumeCollection"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
  /volumes/{name}:
    get:
      operationId: getVolume
      tags: [Volumes]
      summary: Volume detail
      description: |
        Returns the volume and a list of services using it.
        Note: volumes are keyed by name, not ID.
      parameters:
        - name: name
          in: path
          required: true
          description: Volume name
          schema:
            type: string
      responses:
        "200":
          description: Volume detail with cross-referenced services
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/DetailEnvelope"
                  - type: object
                    properties:
                      volume:
                        type: object
                        description: Docker volume.Volume object
                      services:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceRef"
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeVolume
      tags: [Volumes]
      summary: Remove a volume
      description: Removes a volume. Fails if the volume is in use by a container. Requires operations level 3.
      parameters:
        - name: name
          in: path
          required: true
          description: Volume name
          schema:
            type: string
      responses:
        "204":
          description: Volume removed
        "403":
          description: Operations level too low
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Search ───────────────────────────────────────────────────────────
  /search:
    get:
      operationId: search
      tags: [Search]
      summary: Global cross-resource search
      description: |
        Searches names, images, and labels across all resource types.
        Results are grouped by type. The `limit` parameter controls max results
        per type (default 3; 0 means up to 1000 per type).
      parameters:
        - name: q
          in: query
          required: true
          description: Search query (case-insensitive substring match, max 200 chars)
          schema:
            type: string
            maxLength: 200
        - name: limit
          in: query
          description: Max results per resource type (0 = up to 1000)
          schema:
            type: integer
            default: 3
            minimum: 0
      responses:
        "200":
          description: Grouped search results
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Topology ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-cetacean_session
      description: Session cookie set after OIDC authentication
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for machine-to-machine OIDC authentication

  # ── Parameters ───────────────────────────────────────────────────────
  parameters:
    resourceId:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: string
    search:
      name: search
      in: query
      description: Case-insensitive substring search on resource names
      schema:
        type: string
    filter:
      name: filter
      in: query
      description: |
        expr-lang filter expression (max 512 characters).
        Available fields depend on the resource type.
      schema:
        type: string
        maxLength: 512
    limit:
      name: limit
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    offset:
      name: offset
      in: query
      description: Starting offset for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    dir:
      name: dir
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: asc
...
  headers:
    ETag:
      description: Weak ETag for conditional requests (use with If-None-Match)
      schema:
        type: string
        example: '"w/abc123"'
    Link:
      description: RFC 8288 pagination links (rel="next", rel="prev")
      schema:
        type: string
        example: '</nodes?limit=50&offset=50>; rel="next"'
...
  responses:
    NotModified:
      description: Not modified (ETag matched If-None-Match)
    BadRequest:
      description: Bad request (invalid filter, missing parameter, etc.)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Conflict:
      description: Version conflict — the resource was modified concurrently; retry the operation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    UnsupportedMediaType:
      description: Unsupported Content-Type for this endpoint
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    ProblemDetail:
      type: object
      description: |
        RFC 9457 Problem Details with JSON-LD @context.
        Domain-specific errors use `/api/errors/{code}` as the type URI
        (e.g. `/api/errors/SVC001`). Generic errors use `about:blank`.
      required: [type, title, status]
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        type:
          type: string
          description: "Problem type URI — `/api/errors/{code}` for domain errors, `about:blank` for generic"
          example: /api/errors/SVC001
        title:
          type: string
          description: Short human-readable summary
          example: Service Version Conflict
        status:
          type: integer
          description: HTTP status code
          example: 409
        detail:
          type: string
          description: Human-readable explanation
        instance:
          type: string
          description: Request path that generated the error
        requestId:
          type: string
          description: Correlation ID for this request
...
    DetailEnvelope:
      type: object
      description: JSON-LD wrapper for detail endpoints
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@id":
          type: string
          description: Canonical path of this resource
          example: /nodes/abc123
        "@type":
          type: string
          description: Resource type name
          example: Node
    CollectionEnvelope:
      type: object
      description: JSON-LD wrapper for list endpoints
      required: [items, total, limit, offset]
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@type":
          type: string
          example: Collection
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total number of items matching the query
        limit:
          type: integer
          description: Page size used
        offset:
          type: integer
          description: Starting offset used

    # ── Collections ──────────────────────────────────────────────────
...
    ConfigCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Docker swarm.Config object
...
    VolumeCollection:
      allOf:
        - $ref: "#/components/schemas/CollectionEnvelope"
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Docker volume.Volume object
...
    ServiceRef:
      type: object
      description: Lightweight service cross-reference
      properties:
        "@id":
          type: string
          description: Canonical path
          example: /services/abc123
        id:
          type: string
        name:
          type: string
...
    SearchResponse:
      type: object
      properties:
        "@context":
          type: string
          example: /api/context.jsonld
        "@id":
          type: string
          example: /search
        "@type":
          type: string
          example: SearchResult
        query:
          type: string
        results:
          type: object
          description: Grouped results keyed by resource type
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/SearchResult"
        counts:
          type: object
          description: Total match count per type (may exceed returned results)
          additionalProperties:
            type: integer
        total:
          type: integer
          description: Total matches across all types
    SearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        detail:
          type: string
          description: Context-dependent detail (image, driver, date, etc.)
        state:
          type: string
          description: Present for services and tasks

    # ── Topology ─────────────────────────────────────────────────────
...
    JSONPatch:
      type: array
      description: JSON Patch document (RFC 6902)
      items:
        type: object
        required: [op, path]
        properties:
          op:
            type: string
            enum: [add, remove, replace, test]
            description: Patch operation
          path:
            type: string
            description: Key name (with or without leading /)
          value:
            type: string
            description: New value (required for add/replace/test)
...